+++ packages/core/src/types/index.ts
@@ -4,3 +4,4 @@ export * from "./request.types";
 export * from "./schema.types";
 export * from "./client.types";
 export * from "./instance.types";
+export * from "./trigger.types";